refact: interaction improvemnets
