+++ src/Social/Contracts/Follows/FollowableInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kanvas\Packages\Social\Contracts\Follows;
+
+interface FollowableInterface
+{
+    public function getId();
+}
