+++ src/Social/Contracts/Users/UserInterface.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Kanvas\Packages\Social\Contracts\Users;
 
+/**
+ * @deprecated v0.4
+ */
 interface UserInterface
 {
     public function getDefaultCompany();
