+++ src/Social/Enums/Interactions.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace Kanvas\Packages\Social\Enums;
+
+class Interactions
+{
+    const REACT = 'react';
+    const SAVE = 'save';
+    const LIKE = 'like';
+    const SHARE = 'share';
+    const COMMENT = 'comment';
+    const REPLY = 'replY';
+    const FOLLOW = 'follow';
+    const BUY = 'buy';
+}
