+++ tests/_support/Models/Tag.php
@@ -2,8 +2,8 @@
 
 namespace Kanvas\Packages\Test\Support\Models;
 
-use Kanvas\Packages\Social\Contracts\Interactions\FollowableInterface;
-use Kanvas\Packages\Social\Contracts\Interactions\FollowersTrait;
+use Kanvas\Packages\Social\Contracts\Follows\FollowableInterface;
+use Kanvas\Packages\Social\Contracts\Follows\FollowersTrait;
 
 class Tag extends BaseModel implements FollowableInterface
 {
@@ -11,7 +11,7 @@ class Tag extends BaseModel implements FollowableInterface
 
     public int $id = 1;
 
-    public function getId(): int
+    public function getId() : int
     {
         return $this->id;
     }
