+++ tests/_support/Models/Users.php
@@ -2,12 +2,13 @@
 
 namespace Kanvas\Packages\Test\Support\Models;
 
-use Kanvas\Packages\Social\Contracts\Interactions\FollowableInterface;
-use Kanvas\Packages\Social\Contracts\Interactions\FollowersTrait;
-use Baka\Contracts\Auth\UserInterface;
 use Canvas\Models\Users as ModelsUsers;
+use Kanvas\Packages\Social\Contracts\Follows\FollowableInterface;
+use Kanvas\Packages\Social\Contracts\Follows\FollowersTrait;
+use Kanvas\Packages\Social\Contracts\Interactions\InteractionsTrait;
 
 class Users extends ModelsUsers implements FollowableInterface
 {
     use FollowersTrait;
+    use InteractionsTrait;
 }
