+++ tests/integration/Social/Services/TagsCest.php
@@ -3,10 +3,10 @@
 namespace Kanvas\Packages\Tests\Integration\Social\Service;
 
 use IntegrationTester;
-use Kanvas\Packages\Social\Models\Interactions as ModelsInteractions;
+use Kanvas\Packages\Social\Enums\Interactions as EnumsInteractions;
+use Kanvas\Packages\Social\Interactions;
 use Kanvas\Packages\Social\Models\Tags as ModelsTags;
-use Kanvas\Packages\Social\Services\Interactions;
-use Kanvas\Packages\Social\Services\Tags;
+use Kanvas\Packages\Social\Tags;
 use Kanvas\Packages\Test\Support\Models\Users;
 
 class TagsCest
@@ -93,7 +93,7 @@ class TagsCest
     public function tagsInteraction(IntegrationTester $I) : void
     {
         $I->assertFalse(
-            Interactions::add(Users::findFirst(1), $this->tag, ModelsInteractions::FOLLOWING)
+            Interactions::add(Users::findFirst(1), $this->tag, EnumsInteractions::LIKE)
         );
     }
 }
